refactor(learning): remove content field from Lesson aggregate chore(flyway): update initial schema migration feat(media): introduce Media aggregate for content management

+++ src/main/java/com/deutschhub/application/learning/dto/response/AdminCourseEnrollmentResponse.java
@@ -1,5 +1,6 @@
 package com.deutschhub.application.learning.dto.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -10,7 +11,7 @@ public record AdminCourseEnrollmentResponse(
         String status,
         int completedLessons,
         int totalLessons,
-        double completionPercentage,
+        BigDecimal completionPercentage,
         int totalStudyMinutes,
         LocalDateTime enrolledAt,
         LocalDateTime lastProgressUpdatedAt
