+++ src/main/java/com/deutschhub/application/learning/dto/response/EnrollmentProgressResponse.java
@@ -1,5 +1,6 @@
 package com.deutschhub.application.learning.dto.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -9,7 +10,7 @@ public record EnrollmentProgressResponse(
         String status,
         int completedLessons,
         int totalLessons,
-        double completionPercentage,
+        BigDecimal completionPercentage,
         int totalStudyMinutes,
         LocalDateTime lastProgressUpdatedAt
 ) {
