+++ src/main/java/com/deutschhub/application/learning/dto/response/MyCourseResponse.java
@@ -15,7 +15,7 @@ public record MyCourseResponse(
         String enrollmentStatus,
         int completedLessons,
         int totalLessons,
-        double completionPercentage,
+        BigDecimal completionPercentage,
         int totalStudyMinutes,
         LocalDateTime enrolledAt,
         LocalDateTime lastProgressUpdatedAt
